Fixing broken test
