+++ src/WebchatSetting/Service/WebchatSettingService.php
@@ -13,4 +13,34 @@ class WebchatSettingService implements WebchatSettingServiceInterface
             ));
         }
     }
+
+    /**
+     * From: https://gist.github.com/koenhoeijmakers/0a8e326ee3b12a826d73be38693fb647
+     *
+     * @param array $original
+     * @param array $merging
+     * @return array
+     */
+    private function recursiveMergeConfigFrom(array $original, array $merging)
+    {
+        $array = array_merge($original, $merging);
+
+        foreach ($original as $key => $value) {
+            if (!is_array($value)) {
+                continue;
+            }
+
+            if (!Arr::exists($merging, $key)) {
+                continue;
+            }
+
+            if (is_numeric($key)) {
+                continue;
+            }
+
+            $array[$key] = $this->recursiveMergeConfigFrom($value, $merging[$key]);
+        }
+
+        return $array;
+    }
 }
