+++ src/WebchatSetting/WebchatSettingServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace OpenDialogAi\Webchat\WebchatSetting;
 
-use Illuminate\Support\Arr;
 use Illuminate\Support\ServiceProvider;
 use OpenDialogAi\Webchat\WebchatSetting\Service\WebchatSettingService;
 use OpenDialogAi\Webchat\WebchatSetting\Service\WebchatSettingServiceInterface;
@@ -28,35 +27,5 @@ class WebchatSettingServiceProvider extends ServiceProvider
     public function mergeConfigFrom($path, $key)
     {
         resolve(WebchatSettingServiceInterface::class)->mergeConfigFrom($path, $key);
-    }
-
-    /**
-     * From: https://gist.github.com/koenhoeijmakers/0a8e326ee3b12a826d73be38693fb647
-     *
-     * @param array $original
-     * @param array $merging
-     * @return array
-     */
-    private function recursiveMergeConfigFrom(array $original, array $merging)
-    {
-        $array = array_merge($original, $merging);
-
-        foreach ($original as $key => $value) {
-            if (!is_array($value)) {
-                continue;
-            }
-
-            if (!Arr::exists($merging, $key)) {
-                continue;
-            }
-
-            if (is_numeric($key)) {
-                continue;
-            }
-
-            $array[$key] = $this->recursiveMergeConfigFrom($value, $merging[$key]);
-        }
-
-        return $array;
     }
 }
